+++ pranama-ui/src/components/View/services/login_service/LoginForm.vue
@@ -98,7 +98,7 @@
   </div>
 </template>
 <script>
-import ContactService from "@/services/api/ContactService";
+import ContactService from "@/shared/api/ContactService";
 
 export default {
   data() {
